backend/requirements-310.txt
groq
moviepy==1.0.3
yt-dlp
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
requests
imageio-ffmpeg
